# Conflicts:
#	hopsworks-api/src/main/java/io/hops/hopsworks/api/filter/RequestAuthFilter.java
#	hopsworks-api/src/main/java/io/hops/hopsworks/api/project/ProjectService.java
#	hopsworks-api/src/main/java/io/hops/hopsworks/api/user/AuthService.java
#	hopsworks-api/src/main/java/io/hops/hopsworks/rest/application/config/ApplicationConfig.java
#	hopsworks-common/src/main/java/io/hops/hopsworks/common/dao/kafka/KafkaFacade.java
#	hopsworks-web/yo/app/index.html
#	hopsworks-web/yo/app/scripts/controllers/pythonDepsCtrl.js
#	hopsworks-web/yo/app/scripts/services/ModalService.js
#	hopsworks-web/yo/app/views/projectSettings.html

+++ hopsworks-common/src/main/java/io/hops/hopsworks/common/dao/hdfs/inode/InodePK.java
@@ -1,3 +1,42 @@
+/*
+ * Changes to this file committed after and not including commit-id: ccc0d2c5f9a5ac661e60e6eaf138de7889928b8b
+ * are released under the following license:
+ *
+ * This file is part of Hopsworks
+ * Copyright (C) 2018, Logical Clocks AB. All rights reserved
+ *
+ * Hopsworks is free software: you can redistribute it and/or modify it under the terms of
+ * the GNU Affero General Public License as published by the Free Software Foundation,
+ * either version 3 of the License, or (at your option) any later version.
+ *
+ * Hopsworks is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ * PURPOSE.  See the GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License along with this program.
+ * If not, see <https://www.gnu.org/licenses/>.
+ *
+ * Changes to this file committed before and including commit-id: ccc0d2c5f9a5ac661e60e6eaf138de7889928b8b
+ * are released under the following license:
+ *
+ * Copyright (C) 2013 - 2018, Logical Clocks AB and RISE SICS AB. All rights reserved
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this
+ * software and associated documentation files (the "Software"), to deal in the Software
+ * without restriction, including without limitation the rights to use, copy, modify, merge,
+ * publish, distribute, sublicense, and/or sell copies of the Software, and to permit
+ * persons to whom the Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS  OR IMPLIED, INCLUDING
+ * BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL  THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR  OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
 package io.hops.hopsworks.common.dao.hdfs.inode;
 
 import java.io.Serializable;
@@ -13,7 +52,7 @@ public class InodePK implements Serializable {
   @Basic(optional = false)
   @NotNull
   @Column(name = "parent_id")
-  private int parentId;
+  private long parentId;
 
   @Basic(optional = false)
   @NotNull
@@ -25,22 +64,22 @@ public class InodePK implements Serializable {
   @Basic(optional = false)
   @NotNull
   @Column(name = "partition_id")
-  private int partitionId;
+  private long partitionId;
 
   public InodePK() {
   }
 
-  public InodePK(int parentId, String name, int partitionId) {
+  public InodePK(long parentId, String name, long partitionId) {
     this.parentId = parentId;
     this.name = name;
     this.partitionId = partitionId;
   }
 
-  public int getParentId() {
+  public long getParentId() {
     return parentId;
   }
 
-  public void setParentId(int parentId) {
+  public void setParentId(long parentId) {
     this.parentId = parentId;
   }
 
@@ -52,11 +91,11 @@ public class InodePK implements Serializable {
     this.name = name;
   }
 
-  public int getPartitionId() {
+  public long getPartitionId() {
     return partitionId;
   }
 
-  public void setPartitionId(int partitionId) {
+  public void setPartitionId(long partitionId) {
     this.partitionId = partitionId;
   }
